# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Event Space Booking System API Documentation'
  description: 'API for managing event space bookings, calendar events, and staff assignments.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by making a POST request to `/api/v1/login` with your email and password.'
security:
  -
    default: []
paths:
...
  /api/v1/events/check-availability:
    post:
      summary: 'Check Availability'
      operationId: checkAvailability
      description: 'Check if a specific event space is available for the given date range.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                oneOf:
                  -
                    description: 'Space is available'
                    type: object
                    example:
                      success: true
                      data:
                        available: true
                        message: 'Space is available'
                    properties:
                      success:
                        type: boolean
                        example: true
                      data:
                        type: object
                        properties:
                          available:
                            type: boolean
                            example: true
                          message:
                            type: string
                            example: 'Space is available'
                  -
                    description: 'Space has conflicts'
                    type: object
                    example:
                      success: true
                      data:
                        available: false
                        message: 'Space is not available for the selected dates'
                    properties:
                      success:
                        type: boolean
                        example: true
                      data:
                        type: object
                        properties:
                          available:
                            type: boolean
                            example: false
                          message:
                            type: string
                            example: 'Space is not available for the selected dates'
      tags:
        - 'Events & Bookings'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_space_id:
                  type: integer
                  description: 'The ID of the event space to check'
                  example: 1
                start_date:
                  type: date
                  description: 'Start date (Y-m-d)'
                  example: '2024-06-15'
                end_date:
                  type: date
                  description: 'End date (Y-m-d)'
                  example: '2024-06-16'
              required:
                - event_space_id
                - start_date
                - end_date
      security: []
...